+++ src/rocprof_compute_profile/profiler_rocprof_v3.py
@@ -64,8 +64,25 @@ class rocprof_v3_profiler(RocProfCompute_Base):
             trace_option,
             "--output-format",
             rocprof_out_format,
-            "--",
         ]
+        # Kernel filtering
+        if self.get_args().kernel:
+            args.extend(["--kernel-include-regex", "|".join(self.get_args().kernel)])
+        # Dispatch filtering
+        dispatch = []
+        # rocprofv3 dispatch indexing is inclusive and starts from 1
+        if self.get_args().dispatch:
+            for dispatch_id in self.get_args().dispatch:
+                if ":" in dispatch_id:
+                    tokens = dispatch_id.split(":")
+                    # 4:7 -> 5-7
+                    dispatch.append(f"{int(tokens[0]) + 1}-{tokens[1]}")
+                else:
+                    # 4 -> 5
+                    dispatch.append(f"{int(dispatch_id) + 1}")
+        if dispatch:
+            args.extend(["--kernel-iteration-range", f"[{','.join(dispatch)}]"])
+        args.append("--")
         args.extend(app_cmd)
         return args