+++ CHANGELOG.md
@@ -16,10 +16,16 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
 ### Changed
 
 * Change normal_unit default to per_kernel
+* change dependency from rocm-smi to amd-smi
 
 ### Resolved issues
 
 * Fixed option specs-correction
+* Fixed kernel name and kernel dispatch filtering when using rocprof v3
+
+### Known issues
+
+* gpu id filtering is not supported when using rocprof v3
 
 ## (Unreleased) ROCm Compute Profiler 3.1.0 for ROCm 6.4.0